stuffs

+++ omp/main.c
@@ -46,7 +46,7 @@ int main(int argc, char *argv[])
 
 	// Hash original file
 	hash_original = hash(file, numblocks);
-	printf("Original hash = %llx\n", hash_original);
+	printf("Original hash = %llx\n", (unsigned long long)hash_original);
 
 	#pragma omp parallel
 	if (omp_get_thread_num() == 0)
@@ -71,7 +71,7 @@ int main(int argc, char *argv[])
 	printf("Average speed: %lf MB/s\n", (double)filesize/(runtime*MEGABYTE));
 
 	hash_encrypted = hash(file, numblocks);
-	printf("Encrypted hash = %llx\n", hash_encrypted);
+	printf("Encrypted hash = %llx\n", (unsigned long long)hash_encrypted);
 
 	//__________DECRYPTION__________
 	printf("Decryption starts...\n");
@@ -92,7 +92,7 @@ int main(int argc, char *argv[])
 	printf("Average speed: %lf MB/s\n", (double)filesize/(runtime*MEGABYTE));
 
 	hash_decrypted = hash(file, numblocks);
-	printf("Decrypted hash = %llx\n", hash_decrypted);
+	printf("Decrypted hash = %llx\n", (unsigned long long)hash_decrypted);
 
 	// Check
 	if (hash_decrypted == hash_original)
